--- packages/next/src/hooks.ts
@@ -1,2 +1,0 @@
-export * from './action-hooks'
-export * from './safe-action-hooks'
